requirements.txt
pandas>=1.3.0
PyYAML>=5.4.0
tqdm>=4.62.0
openpyxl>=3.0.7
numpy>=1.21.0